requirements.txt
exa_py
flask
python-dotenv
